+++ wayflowcore/src/wayflowcore/serialization/__init__.py
@@ -4,8 +4,14 @@
 # (LICENSE-APACHE or http://www.apache.org/licenses/LICENSE-2.0) or Universal Permissive License
 # (UPL) 1.0 (LICENSE-UPL or https://oss.oracle.com/licenses/upl), at your option.
 
-from typing import Any
-
+from .conversation import (
+    deserialize_conversation,
+    deserialize_conversation_state,
+    dump_conversation_state,
+    dump_variable_state,
+    load_conversation_state,
+    serialize_conversation_state,
+)
 from .serializer import (
     autodeserialize,
     deserialize,
@@ -13,43 +19,6 @@ from .serializer import (
     serialize,
     serialize_to_dict,
 )
-
-
-def dump_conversation_state(*args: Any, **kwargs: Any) -> Any:
-    from .conversation import dump_conversation_state as _dump_conversation_state
-
-    return _dump_conversation_state(*args, **kwargs)
-
-
-def serialize_conversation_state(*args: Any, **kwargs: Any) -> Any:
-    from .conversation import serialize_conversation_state as _serialize_conversation_state
-
-    return _serialize_conversation_state(*args, **kwargs)
-
-
-def deserialize_conversation_state(*args: Any, **kwargs: Any) -> Any:
-    from .conversation import deserialize_conversation_state as _deserialize_conversation_state
-
-    return _deserialize_conversation_state(*args, **kwargs)
-
-
-def load_conversation_state(*args: Any, **kwargs: Any) -> Any:
-    from .conversation import load_conversation_state as _load_conversation_state
-
-    return _load_conversation_state(*args, **kwargs)
-
-
-def deserialize_conversation(*args: Any, **kwargs: Any) -> Any:
-    from .conversation import deserialize_conversation as _deserialize_conversation
-
-    return _deserialize_conversation(*args, **kwargs)
-
-
-def dump_variable_state(*args: Any, **kwargs: Any) -> Any:
-    from .conversation import dump_variable_state as _dump_variable_state
-
-    return _dump_variable_state(*args, **kwargs)
-
 
 __all__ = [
     "autodeserialize",